+++ js/firekue/RESTworker.js
@@ -30,7 +30,7 @@ var URL = require("url");
     RESTworker.prototype.setPort = function(port) {
         var that = this;
         this.defaultPort = port;
-        console.log("RESTworker.setPort(" + port + ")");
+        that.verbose && verboseLooger.debug("RESTworker.setPort(" + port + ")");
     }
     RESTworker.prototype.dataAt = function(iData) {
         var that = this;